requirements.txt
Django==1.6.5
Jinja2==2.7.3
MarkupSafe==0.23
Pygments==1.6
Sphinx==1.2.2
argparse==1.2.1
distribute==0.6.24
django-appconf==0.6
django-bootstrap3==4.7.0
django-compressor==1.4
django-extensions==1.3.8
django-geojson==2.5.0
django-haystack==2.1.0
django-leaflet==0.13.5
django-registration==1.0
django-reversion==1.8.1
docutils==0.11
fake-factory==0.4.0
psycopg2==2.5.3
pygraphviz==1.2
python-ptrace==0.7
six==1.7.2
wsgiref==0.1.2
